+++ src/io/oss_io.ml
@@ -34,7 +34,7 @@ let force f fd x =
     if x <> x' then failwith "cannot obtain desired OSS settings"
 
 (** Dedicated clock. *)
-let get_clock = Tutils.lazy_cell (fun () -> new Clock.self_sync "OSS")
+let get_clock = Tutils.lazy_cell (fun () -> new Clock.wallclock ~sync:true "OSS")
 
 class output ~kind ~clock_safe ~on_start ~on_stop 
              ~infallible ~start dev val_source =